requirements.txt
aioredis==1.3.0
amqp==2.5.2
apturl==0.5.2
asgi-redis==1.4.3
asgiref==3.2.3
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.3.0
Authlib==0.13
autobahn==19.11.1
Automat==0.8.0
awscli==1.16.290
billiard==3.6.1.0
blinker==1.4
boto==2.49.0
boto3==1.10.25
botocore==1.13.28
bottle==0.12.17
bottle-pyjwt==1.3.4.post0
Brlapi==0.6.7
celery==4.3.0
certifi==2018.8.24
cffi==1.13.2
channels==2.3.1
channels-redis==2.4.1
chardet==3.0.4
CodeConvert==3.0.2
colorama==0.4.1
command-not-found==0.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.4
cryptography==2.8
cupshelpers==1.0
daphne==2.4.0
defer==1.0.6
defusedxml==0.6.0
distro==1.3.0
distro-info===0.21ubuntu2
Django==2.0.2
django-admin==2.0.0
django-allauth==0.40.0
django-authlib==0.9.3
django-channels==0.7.0
django-dotenv==1.4.2
django-elasticsearch-dsl==7.1.0
django-elasticsearch-dsl-drf==0.20.3
django-excel-response2==2.0.8
django-extensions==1.9.9
django-filter==1.1.0
django-models-ext==1.1.9
django-nine==0.2.2
django-oauth-toolkit==1.2.0
django-redis==4.10.0
django-redis-pubsub==1.0
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
django-short-url==1.1.4
django-six==1.0.4
django-social-share==1.3.2
django-ssl-auth==2.0.1
django-storages==1.8
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
docutils==0.15.2
elasticsearch==7.1.0
elasticsearch-dsl==7.1.0
entrypoints==0.3
furl==2.1.0
futures==2.2.0
happyjwt==0.2.0
hiredis==0.2.0
httplib2==0.11.3
hyperlink==19.0.0
idna==2.6
importlib-metadata==0.23
incremental==17.5.0
itypes==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
keyring==17.1.1
keyrings.alt==3.1.1
kombu==4.6.6
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.8.0
macaroonbakery==1.2.1
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.1.0
more-itertools==7.2.0
msgpack==0.6.2
msgpack-python==0.5.6
mysqlclient==1.4.6
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
openapi-codec==1.3.2
orderedmultidict==1.0.1
packaging==19.2
pexpect==4.6.0
Pillow==5.4.1
pluggy==0.13.1
privex-pyjwt==1.8.0
protobuf==3.6.1
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.9.4
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.32.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pyjwt-rsa==0.3.0
PyLin==0.2.7
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.5
pyRFC3339==1.1
pyserial==3.4
pytest==5.3.1
pytest-dependency==0.4.0
pytest-django==3.7.0
python-apt==1.8.4
python-dateutil==2.8.0
python-debian==0.1.34
python-dotenv==0.10.3
python3-openid==3.1.0
pytz==2018.3
pyxdg==0.25
PyYAML==5.1.2
redis==2.10.5
reportlab==3.5.18
requests==2.21.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
roman==2.0.0
rsa==3.4.2
s3transfer==0.2.1
screen==1.0.1
SecretStorage==2.3.1
service-identity==18.1.0
shortuuid==0.5.0
simplejson==3.16.0
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
system-service==0.3
systemd-python==234
TimeConvert==1.5.0
Twisted==19.10.0
txaio==18.8.1
typing==3.6.4
tzlocal==2.0.0
ubuntu-advantage-tools==19.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.0
urllib3==1.25.7
usb-creator==0.3.3
vine==1.3.0
virtualenv==15.1.0
wadllib==1.3.3
wcwidth==0.1.7
Werkzeug==0.16.0
xkit==0.0.0
xlwt==1.3.0
zipp==0.6.0
zope.interface==4.7.1
